Input

+++ inc/player.h
@@ -5,8 +5,8 @@ typedef struct s_player
 {
 	int	pos_x;
 	int	pos_y;
-	int	dir_x;
-	int	dir_y;
+	float	dir_x;
+	float	dir_y;
 } t_player;
 
 #endif
